tests/presentation/controllers/survey/LoadSurveysController.spec.ts
import { LoadSurveysController } from '@/presentation/controllers/survey/LoadSurveysController'
import { noContent, ok, serverError } from '@/presentation/helpers/http'
import { mockLoadSurveys, mockSurveys } from '@/../tests/domain/mocks'
import type { LoadSurveysControllerRequest } from '@/presentation/controllers/survey/LoadSurveysController'
import type { LoadSurveys } from '@/domain/useCases/survey/LoadSurveys'

type Sut = {
  sut: LoadSurveysController
  loadSurveysStub: LoadSurveys
}

const makeSut = (): Sut => {
  const loadSurveysStub = mockLoadSurveys()
  const sut = new LoadSurveysController(loadSurveysStub)
  return {
    sut,
    loadSurveysStub
  }
}

const mockRequest = (): LoadSurveysControllerRequest => ({
  accountId: 'account_id'
})

describe('Load Surveys Controller', () => {
  test('Should call LoudSurveys with correct value', async () => {
    const { sut, loadSurveysStub } = makeSut()
    const loadSpy = jest.spyOn(loadSurveysStub, 'load')
    const request = mockRequest()
    await sut.handle(request)
    expect(loadSpy).toHaveBeenCalledWith(request.accountId)
  })

  test('Should return 200 on success', async () => {
    const { sut } = makeSut()
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(ok(mockSurveys()))
  })

  test('Should return 204 if LoadSurveys returns an empty array', async () => {
    const { sut, loadSurveysStub } = makeSut()
    jest.spyOn(loadSurveysStub, 'load').mockResolvedValueOnce([])
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(noContent())
  })

  test('Should return 500 if LoudSurveys throws', async () => {
    const { sut, loadSurveysStub } = makeSut()
    jest.spyOn(loadSurveysStub, 'load').mockImplementationOnce(() => { throw new Error() })
    const httpResponse = await sut.handle(mockRequest())
    expect(httpResponse).toEqual(serverError(new Error()))
  })
})
